Brandon v.0.3.1 | UI OVERHAUL

+++ app/src/main/java/com/example/lusonus/ui/composables/MediaLibraryComposables/MediaLibraryItem.kt
@@ -1,12 +1,28 @@
 package com.example.lusonus.ui.composables.MediaLibraryComposables
 
+import FileRow
 import android.net.Uri
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.dp
@@ -18,33 +34,55 @@ fun MediaLibraryItem(
     onDelete: (Uri) -> Unit,
     onClick: (String) -> Unit,
 ) {
-    Surface(
-        color = MaterialTheme.colorScheme.surfaceVariant,
-        contentColor = MaterialTheme.colorScheme.onSurfaceVariant,
+    var deleteMode by rememberSaveable { mutableStateOf(false) }
+
+    val containerColor =
+        if (deleteMode) MaterialTheme.colorScheme.error
+        else MaterialTheme.colorScheme.surfaceVariant
+
+    Card(
+        colors = CardDefaults.cardColors(
+            containerColor = containerColor
+        ),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
         shape = MaterialTheme.shapes.medium,
-        shadowElevation = 4.dp,
         modifier = Modifier
             .fillMaxWidth()
-            .padding(vertical = 4.dp, horizontal = 8.dp)
-            // THIS IS SO COOL! It's how you can check for long presses!
-            .pointerInput(media.uri) {
-                detectTapGestures(
-                    // We specify a long press.
-                    onLongPress = {
-                        onDelete(media.uri)
-                    },
-
-                    // AND THEN WE CAN JUST HAVE... ITS AMAZING!
-                    onTap = {
-
-                        onClick(media.name)
-                    }
-                )
-            }
+            .padding(bottom = 16.dp)
     ) {
-        FileRow(
-            uri = media.uri,
-            modifier = Modifier.padding(12.dp)
-        )
+        Box(
+            modifier = Modifier
+                .combinedClickable(
+                    onClick = {
+                        if (deleteMode) {
+                            onDelete(media.uri)
+                            deleteMode = !deleteMode
+                        }
+                        else {
+                            onClick(media.name)
+                        }
+                    },
+                    onLongClick = { deleteMode = !deleteMode }
+                ).padding(8.dp)
+        ) {
+            if (deleteMode)
+            {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(64.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Delete,
+                        contentDescription = "Delete",
+                        tint = MaterialTheme.colorScheme.onError,
+                        modifier = Modifier.size(48.dp)
+                    )
+                }
+            } else {
+                FileRow(uri = media.uri)
+            }
+        }
     }
 }
