+++ app/src/main/java/com/example/lusonus/ui/composables/PlaylistLibraryComposables/PlaylistItem.kt
@@ -1,13 +1,27 @@
 package com.example.lusonus.ui.composables.PlaylistLibraryComposables
 
+import FileRow
+import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
@@ -19,36 +33,68 @@ fun PlaylistItem(
     onDelete: (String) -> Unit,
     onClick: (String) -> Unit
 ) {
-    Surface(
-        color = MaterialTheme.colorScheme.secondaryContainer,
-        contentColor = MaterialTheme.colorScheme.secondary,
+    var deleteMode by rememberSaveable { mutableStateOf(false) }
+
+    val containerColor =
+        if (deleteMode) MaterialTheme.colorScheme.error
+        else MaterialTheme.colorScheme.surfaceVariant
+
+    Card(
+        colors = CardDefaults.cardColors(
+            containerColor = containerColor
+        ),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp),
         shape = MaterialTheme.shapes.medium,
-        shadowElevation = 4.dp,
         modifier = Modifier
             .fillMaxWidth()
-            .padding(vertical = 4.dp, horizontal = 8.dp)
-            // THIS IS SO COOL! It's how you can check for long presses!
-            .pointerInput(playlistName) {
-                detectTapGestures(
-                    // We specify a long press.
-                    onLongPress = {
-                        onDelete(playlistName)
-                    },
-
-                    // AND THEN WE CAN JUST HAVE... ITS AMAZING!
-                    onTap = {
-                        onClick(playlistName)
-                    }
-                )
-            }
+            .padding(bottom = 16.dp)
     ) {
         Column(
             modifier = Modifier
+                .combinedClickable(
+                    onClick = {
+                        if (deleteMode) {
+                            onDelete(playlistName)
+                            deleteMode = !deleteMode
+                        }
+                        else {
+                            onClick(playlistName)
+                        }
+                    },
+                    onLongClick = { deleteMode = !deleteMode }
+                )
                 .padding(12.dp)
                 .fillMaxWidth(),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Text(text = playlistName, style = MaterialTheme.typography.bodyLarge)
+            if (deleteMode)
+            {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(56.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Delete,
+                        contentDescription = "Delete",
+                        tint = MaterialTheme.colorScheme.onError,
+                        modifier = Modifier.size(48.dp)
+                    )
+                }
+            } else {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(56.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Text(
+                        text = playlistName,
+                        style = MaterialTheme.typography.bodyLarge
+                    )
+                }
+            }
         }
     }
 }
